syntax: [refactor] Move data types to separate type file
