+++ src/@types/syntax/data.d.ts
@@ -1,0 +1,5 @@
+/** Data-type of a value. */
+export type TData = number | string | boolean;
+
+/** Data-type name of a value. */
+export type TDataName = 'number' | 'string' | 'boolean';
